Use double here
Source Serif Pro dollar glyphs hits a bug where for some numbers the
comparison here would return false on 32-bit arch and true on 64-bit
leading to a missing “hstem” in 32-bit. So this is an attempt to
increase the precision of the numbers so that the result is more
predictable, or so I hope.