Lets bite the bullet and reformat everything
The existing code uses a mix of tab and spaces and different identation
styles that makes it often hard ti follow. Reformatted everything with
clang-format and the included source/.clang-format file. I used Mozilla
style because it looked closer ti the existing code style, but I have no
particular attachment to it.

If one is doing git blame in this commit shows up, you can ask git to
blame the commits before this:
    git blame foo.c this_commit^