Fixed bugs in applying MM hinting.
The problem was that WriteUnmergedHints() was setting outer edge of a
hint to be a delta from the inner edge for all the main hints in all the
masters - but WriteUnmergedHints() was called for each master, so the on
the second pass, the outer edge was getting set to a delta from the
delta. Also, WriteUnmergedHints() was writing only the hints from the
first master for each master.

I also restored the call to WriteUnmergedHints() when the path
elt.hints is not NULL while processing the path elts. 'autohint' does
generate hints in the middle of a path - this is called hint
substitution. The only thing wrong is that hint statements needs to be
enclosed by the keywords for starting and ending hint substitution. You
can see these in the function charpath.c::WriteHints():'beginsubr snc'
and 'endsubr enc\nnewcolors'.

I added documentation about how MM hinting works in the source code.