Use a bit simpler markdown syntax
Those `sh` markers are really superfluous since there is not that much to
syntax highlight anyway, so better keep the file nicely formatted when
read as plain text (that is the appeal of markdown after all). This is
closer to how the it looked before 3f0b4de7f2b7ec5d42062699ac747c2485af9836.