setup multibuild for compiling multi-platform wheels (#55)
* Add multibuild submodule (master branch)

* remove redundant tests/.coveragerc
the root .coveragerc is enough

* .gitignore: we store coverage files in .tox, already ignored

* .gitignore: add htmlcov/ folder

* tests/__init__.py: define DATA_DIR global path to 'tests/data'

always relative to the current working directory, so we can run test modules from anywhere

* test_hintfonts: use DATA_DIR global; save output to tmpdir

* test_hintfontsmm: reference and outputs in tmpdir, use DATA_DIR

py.path is part of pytest (tmpdir fixture itself returns a py.path.local object)

* .coveragerc: fix source/paths, skip_covered, etc.

remove parallel=True as it conficts with pytest-cov and pytest-xdist;
add .tox site-packages folders to 'paths' for combining coverage.

* setup.py: add test requirements to 'testing' extra

so one can install them, e.g.: pip install -e .[testing]

* dev-requirements.txt: only tox; the rest is handled by tox

* tox.ini: run in parallel with combined coverage, etc.

Several changes, including:

- we now use pytest-xdist plugin to distribute tests across several
  processes, depending on the CPU cores (-n auto)
- py37 added to default envlist (but skipped if not present on local
  developer's machine)
- pytest now searches in 'testpaths = tests' when no input is given;
  runs in verbose mode, with a summary report at the end
- coverage data files for each tox environment are saved inside the
  .tox work dir, then are combined by a specific 'coverage' tox
  environment which must be run at the end, and which combines and
  produces reports in different formats: console, xml (for codecov),
  and html (saved in 'htmlcov') for inspecting locally in a browser.

* tox.ini: split 'sdist' and 'bdist' environments

we only need to run tox -e sdist on the CI before uploading to PyPI.
I keep 'tox -e bdist' in case one may want to create the wheel locally

* travis: set up multibuild

for now only 2.7 and 3.6 (64-bit) for both linux (wide unicode) and mac.
No deploy stage yet.

* add multibuild's config.sh setup file

run_tests will search for the precompiled psautohint wheel inside 'wheelhouse' dir
and call tox with --installpkg (to skip rebuilding from sdist).

* tox.ini: set TOXPYTHON env variable for finding Appveyor pythons

* tox.ini: add 'wheel' and 'pypi' commands

* appveyor: build wheels for both 32 and 64 bit, etc.

* tox.ini: passenv TWINE_USERNAME and TWINE_PASSWORD

* tox.ini: must define 'basepython' with TOXPYTHON else Appveyor fails

need to tell tox where to find the path to the python interpreter on Windows (do I?)

* config.sh: apply codacy suggestions

https://app.codacy.com/app/frankrolf/psautohint/pullRequest?prid=1866893

* appveyor doesn't like my batch if statement.. let's try powershell

* .travis: need to install tox on Travis default python to be able to use it ourside docker container

* tox.ini: need to create sdist to build wheel from it!

* tox.ini: make 'pypi-wheel' environment; called by Appveyor on tags

* .travis: on osx we need to bootstrap pip in order to install tox...

* travis: move the osx pip bootstrapping in 'after_success'

installing pip seems to break multibuild osx 2.7 build somehow.
https://travis-ci.org/adobe-type-tools/psautohint/jobs/400583973

* appveyor: disable 2.7 64-bit and 3.6 32-bit.. too slow :(

* config.sh: remove .tox folder after run_tests to fix permission errors

the linux builds fail because they can't read/write the .tox folder..

* tox.ini: allow to control PYTEST_NUM_PROCESSES via env variable

* psautohint: always use utf-8 for reading/writing text files

When the locale is not set to UTF-8 (like in the manylinux1 docker container,
where it's set to 'LANG=C'), some of tests fail. Also, it's safe to assume
that UFO xml and plist are alwyas utf-8.

https://travis-ci.org/adobe-type-tools/psautohint/jobs/400901533#L663

* config.sh: .tox dir to clean up is one dir above

run_tests is run from a new temp directory, but .tox dir is always
created next to the tox.ini file, in the parent directory in this case

* ufoFont: fix TypeError when writeHashMap

* travis: after 'before_install', we've already sourced a venv

* ufoFont: use ET.parse(path).getroot() instead of ET.XML(string)

we already have a source filename; we let elementtree library deal with decoding
the text. The XML() factory is for when already starts from a string.
And on py2, this must be a byte string... henche this unicodeencodeerror when
we attempt to pass a unicode string with non-ascii characters:

https://ci.appveyor.com/project/adobe-type-tools/psautohint/build/1.0.237/job/7a4501mbwx2rmqog#L162

* config.in: cd.. back to repository root before calling tox

* tox: use a single 'codecov' env that inherits from 'coverage' env

* put .coverage files outside .tox dir at root of repo

this is because we remove the whole .tox dir after multibuild docker
has completed to avoid permission errors, but then coverage script
cannot find the .coverage files

* travis: add deploy stage; save pip cache

* travis: add encrypted pypi password

* appveyor: add same TWINE_USERNAME and TWINE_PASSWORD as afdko repo

I think these should work...