setup.py: use setuptools instead of meson to build autothintexe
Adding meson as a required build requirement to install psautohint
from source is too heavy.
Meson requires python3 so it cannot be specified like the other python
build requirements (e.g. setuptools-scm) in the pyproject.toml, because
we still care about supporting python2.7 for a while.

If we are happy to compile the libpsautohint C static library using
setuptools and use it with the psautohint extension module, then we
should be fine with using Distutils' CCompiler.link_executable to
create autohintexe.

Also, the latter needs not be installed in the Python's bin or
Scripts, because it is not a user-facing script (unlike, say, tx),
but can only be used with the old afdko's autohint.py wrapper.
There's no need to hack distutils to install autohintexe alongside
other text-only 'scripts'.
We can distribute autohintexe inside the psautohint package and clients
who care about running it as subprocess can get it's filename relative
to `psautohint.__file__` (we are not zip_safe anyway).