tox: don't set basepython to %TOXPYTHON% on Windows
We can simply call 'tox -e py' which means create a tox env with the current
'python' interpreter in the $PATH.
This also allows one to call tox with a python version that is not explicitly
listed in the [testenv] basepython list.
For example, I needed to run tests on python3.5 locally and I couldn't unless
I explicitly listed it in there.
That TOXPYTHON thing was a hack, this is the recommended way to use tox on
windows.