Fix bug in charpath.c that dropped [hv]stem3 hints from MM fonts if any of the source fonts don't met the criteria. This can be relaxed for MM fonts, as these are now hinted as such only for OTF fonts, which use cntrmasks instead, which do not require that all the counters be equal width.
Fix bug in check.c which dropped zero-length paths A path may be zero length in one source font, but not others.

Added to code to 1) apply hint masks and counter masks from default font to other fonts 2) remove hints that are not valid in all fonts, and 3) correct hint and counter masks when invalid hints are removed.

Refactor code to simplify logic, and changed variable names in new code from camel case to snake casing.