Refactor code to make it easier to follow logic for regular vs MM font data. No functional change.
For regular fonts, the logic steps through each font in the list, and hints each glyph in that font. For MM fonts it is simplifies the logic to to step through the glyph list from the default font, and for each glyph in turn, hint the default font glyph and then apply its hints to the matching glyph from the other fonts.