CFF2 VF hinting
Refactored code to allow sharing paths between hinting source fonts for MM fonts, and hinting a CFF2 variable font. Hinting the master sources for a glyph for the two cases is now identical. The latter differs from the former in that the VF charstring must first be interpolated to a regular T2 charstring for each master source font. After the hinting is done, the T2 charstrings must be merged back together into a single VF charstring.

I also renamed some variables that had different meanings for the same name. For example, 'width' in some contexts contains the advance width for a glyph. In other contexts, it contains the T2 charstring width arg, which is (advance width - CFF nominal width). I renamed all occurrences of 'width' in the latter context to 't2_width_arg'.