Fix for issue: Compatible hinting (-r) produces corrupt hinting #189
The problem was that in some cases, psautohint was incorrectly identifying hints as being bad and dropping them in VF hinting. The bug was in the check that all hints were sorted in increasing value of the bottom edge. The bug is that it was comparing the lowest value of successive pairs after converting the top and bottom value of each hint to absolute values.  This can fail with top edge hints, where the delta is the arbitrary "-20": the absolute value of this can be below the bottom value of the previous pair, but this is not a real conflict. Fix this by treating the top edge/bottom edge deltas as being 0, for the purpose of checking sort order and overlap.